refactor: :building_construction: organize better project
