+++ .gitignore
@@ -137,3 +137,5 @@ dist
 
 
 **/*/database/vector/*
+
+**/*/dist/
