+++ src/domain/core/types/dto.ts
@@ -1,0 +1,9 @@
+import type { QuestionAgentDTO } from "../../question/models";
+import type { AgentSourceType } from "./source";
+
+export interface MultiAgentDTO {
+	input: string,
+	llMOutput: QuestionAgentDTO,
+	searchedSources: AgentSourceType[],
+	error?: string
+}
