+++ src/domain/search/selfAskWithSearch/types/permissions.ts
