+++ src/domain/search/tools/getPage/type.ts
