+++ src/domain/search/tools/type.ts
