+++ src/infra/repository/vector.repository.ts
@@ -1,5 +1,5 @@
 import { HNSWLib } from "@langchain/community/vectorstores/hnswlib";
-import type { IVectorStore, VectorDocument, VectorSearchResult } from "../interfaces/vector.repository";
+import type { IVectorStore, VectorDocument, VectorSearchResult } from "@/src/infra/interfaces/vector.repository";
 import type { Embeddings } from "@langchain/core/embeddings";
 import { promises as fs } from "fs";
 import type { SpaceName } from "hnswlib-node";
