+++ src/tools/metadataSQLDatabase.ts
@@ -5,8 +5,8 @@ import { SQL_DATABASE_PATH } from "../config";
 import { AgentLLMService } from "../infra/gateway/agentLlm.service";
 import { ChatPromptTemplate } from "@langchain/core/prompts";
 import { toJsonSchema } from "@langchain/core/utils/json_schema";
-import { databaseMetadataSchema } from "../core/search/tools/musicDB/databaseMetadata";
 import { logger } from "./logger";
+import { databaseMetadataSchema } from "../domain/search/tools/musicDB/databaseMetadata";
 
 function generateMetadata(dbPath: string, sampleSize = 5) {
     if (!fs.existsSync(dbPath)) {
